doc: typeset dummy URLs and protocols in monospace
Wrap dummy URLs and protocols names with backticks as indicated in the
CodingGuidelines. Note that some asciidoc predefined attributes are
replaced by their corresponding characters as they are made literal:

    {tilde} => ~
    {startsb} => [
    {endsb} => ]

Remove backticks of two URLs as they prevent them to be clickable.
Reflect this in the CodingGuidelines: only dummy URLs should be wrapped
with backticks.

Signed-off-by: Firmin Martin <firminmartin24@gmail.com>